requirements.txt
absl-py==2.1.0
acres==0.5.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.8.2
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==45.0.5
curl_cffi==0.11.4
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email-validator==2.3.0
etelemetry==0.3.1
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
filelock==3.20.0
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.59.1
frozendict==2.4.6
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.5.0
kiwisolver==1.4.9
libclang==18.1.1
looseversion==1.3.0
lxml==6.0.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
multitasking==0.0.11
namex==0.0.8
narwhals==1.45.0
networkx==3.5
nibabel==5.3.3
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
openai==1.93.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.11.3
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.1
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.3.8
plotly==6.3.0
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
puremagic==1.30
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
Pygments==2.18.0
PyMuPDF==1.26.3
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.2
qrcode==8.2
rdflib==7.5.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
rich-toolkit==0.15.1
rignore==0.6.4
roboflow==1.2.6
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.36.0
setuptools==74.0.0
shellingham==1.5.4
simplejson==3.20.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.45
sqlmodel==0.0.27
starlette==0.47.3
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.1
tqdm==4.67.1
traits==7.0.2
transformers==4.57.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yfinance==0.2.64
